\name{GofCens-package}
\alias{GofCens-package}
\alias{GofCens}
\docType{package}
\title{Goodness-of-Fit Methods for Right-Censored Data.}
\description{This package implements both graphical tools and goodness-of-fit 
             tests for right-censored data. It has implemented 
\enumerate{
  \item Kolmogorov-Smirnov, Crámer-von Mises, and Anderson-Darling tests based 
  on the empirical distribution function for complete data and their extensions 
  for right-censored data. 
      
  \item Generalized chi-squared-type tests  based on the squared difference between 
  observed and expected counts using random cells with right-censored data.
      
  \item A series of graphical tools such as probability or cumulative hazard 
  plots to guide the decision about the parametric model that best fits the data.
}
}
\details{\tabular{ll}{
    Package: \tab GofCens\cr
    Type: \tab Package\cr
    Version: \tab 0.92\cr
    Date: \tab 2022-01-05\cr
    License: \tab GPL (>= 2)\cr}
}
\author{Klaus Langohr, Mireia Besalú, Guadalupe Gómez

Maintainer: Klaus Langohr <klaus.langohr@upc.edu>}
